+++ app/src/main/java/com/charan/habitdiary/presentation/habits/HabitScreenEffect.kt
@@ -1,9 +1,9 @@
 package com.charan.habitdiary.presentation.habits
 
 sealed class HabitScreenEffect {
-    data class OnNavigateToAddHabitScreen(val id : Int?) : HabitScreenEffect()
+    data class OnNavigateToAddHabitScreen(val id : Long?) : HabitScreenEffect()
 
-    data class OnNavigateToAddDailyLogScreen(val id : Int?) : HabitScreenEffect()
+    data class OnNavigateToAddDailyLogScreen(val id : Long?) : HabitScreenEffect()
 
-    data class OnNavigateToHabitStatsScreen(val habitId : Int) : HabitScreenEffect()
+    data class OnNavigateToHabitStatsScreen(val habitId : Long) : HabitScreenEffect()
 }
